requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.19.0
python-dotenv>=1.0.0
openai>=1.60.0
